feat(utils): separated geocoding from utils
